backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
groq==0.13.1
pydantic==2.10.4
python-multipart==0.0.20
certifi>=2024.2.2
httpx>=0.27.0
